NIC Option read reimplemented for PRO and ELITE
This changeset reverts the changes needed to get option set back into shape after the WIP patch for the elite went in.

The option names are now consitent for PRO and ELITE on the first NIC and matching with the documented names. The second NIC on ELITE devices get options suffixed with a 2.

The Option setting for the Elite is still waiting for deeper testing, will push it if testing went good but i have to wait until we can shedule the downtime we need for the tests.